My Screen Changes
Updating to my branch before merging with primary. Primary is where we can merge to see if our code goes together well and then we put on Master from there. This branch is just to store my code before comparing on primary so that way I can have full access to my changes.