Cleaning up some code and finishing listview additions
ListView is implemented into all seven of the classes it needs to be. SetList doesn't have a ListView so it wasn't given List displaying functionality. Some code formating was also cleaned up a little and some comments were left for methods relating to screen interactions that result in an interaction with the database. Some of those methods were given a vague comment as to what will be done in them.